Refactor: refactoer and clean the code (crypto)
