+++ crypto-app/src/app/auth/signIn/page.tsx
@@ -1,5 +1,5 @@
-import SignInForm from "@/components//ui/SignInForm"
-import LogoutButton from "@/components/ui/LogoutButton"
+import SignInForm from "@/components/auth/SignInForm"
+import LogoutButton from "@/components/auth/LogoutButton"
 import { authOptions, isSession } from "@/utils/auth"
 import { redirect } from "next/navigation"
 import {SignInPage} from "./SignInPage"
