+++ crypto-app/src/app/auth/signUp/SignUpPage.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from "@/components/ui/SignUpForm"
+import SignUpForm from "@/components/auth/SignUpForm"
 
 export const SignUpPage = () => {
     return (
