+++ crypto-app/src/app/crypto/convert/page.tsx
@@ -1,10 +1,10 @@
-import { ConvertCryptoPage } from "./ConvertCryptoPage";
+import { ConvertCryptoSection } from "@/components/dashborad/crypto/ConvertCryptoSection";
 
 
 export default async function Page() {
     return (
         <>
-        <ConvertCryptoPage />
+        <ConvertCryptoSection />
         </>
     )
