+++ crypto-app/src/components/auth/SignUpForm.tsx
