+++ crypto-app/src/components/dashborad/crypto/CryptosRealTimeTable.tsx
@@ -13,8 +13,6 @@ import {
   Dialog,
   DialogContent,
   DialogDescription,
-  DialogHeader,
-  DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog"
 import { Input } from "@/components/ui/input";
@@ -22,10 +20,8 @@ import { trpcClientComp } from "@/server/trpc.Provider";
 import { toast } from "sonner";
 import { useSession } from "next-auth/react";
 import { Button } from "@/components/ui/button";
-import { formatNumber, formatPercentChange, formatTimestamp, getPriceChangeStatus } from "./utils";
-import { SelectCrypto } from "../SelectCrypto";
-import cryptoStore from "@/stores/cryptoStore";
-import CryptoChart from "../CryptoChart";
+import { formatNumber, formatPercentChange, formatTimestamp, getPriceChangeStatus } from "@/utils/client/crypto/utils";
+import CryptoChart from "./CryptoChart";
 
 // Clean interface that matches WebSocket data structure
 interface Crypto {
